pyproject.toml
[tool.poetry]
name = "hybridrag"
version = "0.1.0"
description = ""
authors = ["Abdullah alzariqi"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
langchain = "^0.3.10"
langchain-google-genai = "^2.0.6"
black = "^24.10.0"
isort = "^5.13.2"
langchain-pinecone = "^0.2.0"
langchain-voyageai = "^0.1.3"
python-dotenv = "^1.0.1"
datasets = "^3.1.0"
transformers = "^4.47.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
